backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
sqlalchemy
pydantic
PyJWT
alembic
pytest
